requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-rest-framework==0.1.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
jax==0.4.12
joblib==1.2.0
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1rc1
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.5
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pooch==1.6.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
